% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFitExtendedG.R
\name{GetFitExtendedG}
\alias{GetFitExtendedG}
\title{Compute the extended Fisher's g test statistic}
\usage{
GetFitExtendedG(y)
}
\arguments{
\item{y}{Vector containing the series.}
}
\value{
Vector of length 2 containing the modified Fisher's G test statistic 
and the frequency for the maximum periodogram.
}
\description{
The extended Fisher's g te  statistic is computed for testing for periodicity.
}
\details{
Extend the Fisher's g test by enlarging the searching region 
of the frequency from the fourier frequencies to be 
\eqn{En = {j/101 | j=1,\dots,50 and j/101 \ge 1/n}}.
}
\author{
Yuanhao Lai
}
\references{
Fisher, R.A. (1929). Tests of significance in harmonic analysis. 
Proc. Roy. Soc. A, 125, 54-59.
}
\keyword{internal}

