% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFitFisherG.R
\name{GetFitFisherG}
\alias{GetFitFisherG}
\title{Compute the Fisher's g test statistic}
\usage{
GetFitFisherG(z)
}
\arguments{
\item{z}{Vector containing the series.}
}
\value{
Vector of length 2 containing the modified Fisher's G test statistic 
and the frequency for the maximum periodogram.
}
\description{
The Fisher's g test statistic is computed.
}
\author{
Yuanhao Lai
}
\references{
Fisher, R.A. (1929). Tests of significance in harmonic analysis. 
Proc. Roy. Soc. A, 125, 54-59.
}
\keyword{internal}

