\name{GetFitHReg}
\alias{GetFitHReg}

\title{Compute loglikelihood ratio test statistic }

\description{
The loglikelihood ratio test statistic is computed for testing for periodicity
}

\usage{
GetFitHReg(y, t)
}
\arguments{
  \item{y}{ vector containing the series }
  \item{t}{ vector of corresponding time points }
}
\details{

}
\value{
vector of length 2 containing the residual sum of squares and
estimated frequency
}

\references{  }
\author{  }


\examples{
}

\keyword{ ts }
