\name{GetFitRankG}
\alias{GetFitRankG}

\title{Compute the Robust g test statistic }

\description{
The Robust g test statistic is computed for testing for periodicity.
}

\usage{
GetFitRankG(y)
}
\arguments{
  \item{y}{ vector containing the series. }
}
\details{

}
\value{
The Robust g test statistic. 
}

\references{ 
 Ahdesmaki, M., Lahdesmaki, H., Pearson, R., Huttunen, H., and 
Yli-Harja O.
  (2005).  \emph{BMC Bioinformatics} \bold{6}:117. \url{http://www.biomedcentral.com/1471-2105/6/117}
}
\author{ Yuanhao Lai. }


\examples{
}

\keyword{ ts }
