% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FisherGTest.R
\name{FisherGTest}
\alias{FisherGTest}
\title{Fisher's g test}
\usage{
FisherGTest(z)
}
\arguments{
\item{z}{Input series, no missing values.}
}
\value{
Vector of length 3 containing the g test statistic, the p-value, 
and the frequency for the maximum periodogram.
}
\description{
Calculates p-value using Fisher's g-test for a single frequency.
}
\details{
Program is interfaced to C for efficient computation.
}
\author{
A.I. McLeod
}
\references{
Fisher, R.A. (1929). Tests of significance in harmonic analysis. 
Proc. Roy. Soc. A, 125, 54-59.
}
\keyword{internal}

