% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFitHReg.R
\name{GetFitHReg}
\alias{GetFitHReg}
\title{Compute loglikelihood ratio test statistic}
\usage{
GetFitHReg(y, t = 1:length(y))
}
\arguments{
\item{t}{Vector of corresponding time points.}

\item{z}{Vector containing the series.}
}
\value{
Vector of length 2 containing -2 loglikelihood ratio 
test statistic and the estimated frequency.
}
\description{
The -2 loglikelihood ratio test statistic, -2LLR, is computed 
for testing for periodicity.
}
\details{
To compute the likelihood ratio statistic, 
a harmonic regression model is fitted by 
the maximum likelihood estimations (MLE)
according to the selected model. In particular, the frequency f is found by 
the grid search among 
\eqn{En = {j/101 | j=1,\dots,50 and j/101 \ge 1/n}}.
The MLE is equivalent to 
the least square estimation.
}
\author{
A.I. McLeod
}
\references{
Islam, M.S. (2008).  Peridocity, 
Change Detection and Prediction in Microarrays. 
Ph.D. Thesis, The University of Western Ontario.
}
\keyword{internal}

